feat: enhance ErrorLoggingService with batch operations and caching
- Add batch operations for better Firestore performance (50 logs per batch)
- Implement caching system with automatic flushing every 30 seconds
- Add separate caches for errors, warnings, and performance logs
- Improve error handling with proper variable scoping
- Add flushAllLogs() method for forced log submission
- Add getLogStats() method for monitoring log statistics
- Add dispose() method for proper resource cleanup
- Create test script for ErrorLoggingService validation
- Maintain backward compatibility with existing API

+++ lib/widgets/review_widgets.dart
@@ -39,7 +39,7 @@ class ReviewCard extends StatelessWidget {
                         ),
                       ),
                     ),
-                    _buildRatingStars(review.rating.toDouble()),
+                    _buildRatingStars(review.rating),
                   ],
                 ),
